First draft of an INSERT guide
This guide mirrors the form of "All About Updates", but covers the
various ways to use `insert_into`.

I opted not to cover upsert in depth, as really PG upsert should have
its own guide if we feel that is useful to have. The API docs are really
comprehensive for upsert, so I think just linking to the docs is fine
for now.

Right now I only included code examples for PG, I'll port that to the
other backends once the guide is ready to be published.